+++ CHANGELOG.md
@@ -2,12 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+### [0.5.0] - 2025-11-03 - Combinatorics Expansions - [ONGOING]
+- Added combinatorics functions:
+  - `isTriangular(n)` – Checks if a number n is triangular
+
+
+
 ### [0.4.1] - 2025-11-03 - Combinatorics Expansions
 - Merged sequence functions into combinatorics
 - Added combinatorics functions:
   - `triangular(n)` – Computes the n-th triangular number
-
-
 
 ### [0.4.0] - 2025-11-02 - Function Expansions